src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
courlan==1.3.2
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
dateparser==1.2.2
distro==1.9.0
curl-cffi>=0.6.2
duckduckgo-search==6.4.2
feedparser==6.0.12
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.2
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-google-genai==4.1.2
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
lxml==6.0.2
lxml_html_clean==0.4.3
mailjet-rest==1.5.1
marshmallow==3.26.2
more-itertools==10.8.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
premailer==3.10.0
primp==0.15.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
serpapi==0.1.5
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
tenacity==9.1.2
tld==0.13.1
trafilatura==2.0.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.2
uuid_utils==0.12.0
websockets==15.0.1
wheel==0.45.1
xxhash==3.6.0
yagmail==0.15.293
yarl==1.22.0
zstandard==0.25.0
